+++ src/node_instance.rs
@@ -99,7 +99,7 @@ pub struct NodeInstanceInfo {
     pub rewards: Option<U256>,
     pub records: Option<usize>,
     pub relevant_records: Option<usize>,
-    pub mem_used: Option<u64>,
+    pub mem_used: Option<f64>,
     pub cpu_usage: Option<String>,
     pub connected_peers: Option<usize>,
     pub kbuckets_peers: Option<usize>,
